Added 2D character animation mechanism.
- Controls (General)
-- Added manual input processing and drawing support for controls.
-- Renamed the methods for enabling/disabling auto processing of controls.
-- Now controls have group IDs to form groups.
-- Fixed the problem that hidden/disabled controls were processed.
-- Changed alpha value of disabled controls: 0.4 -> 0.1.

- Controls (Label)
-- Now KRLabel instances will not accept user inputs.
-- Changed default text alignment of KRLabel.
-- Added text shadow support for KRLabel.

- Controls (Button)
-- Now button texture images won't be drawn in 3-separated way.

- Image Drawing
-- Fixed the problem that particle systems had generate counts of zero by default.
-- Fixed the problem that the drawings of particle systems changed blend mode.
-- Fixed the document description about primitive drawings.

- Debugging
-- Added KRDebug() and KRDebugScreen() functions.

- Misc
-- Added KRMemoryAllocator class to the framework.
-- Added operator<() to KRVector** classes for supporting std::set.
-- Refactored by renaming some file names and class names.

+++ Karakuri Library/Karakuri/KRPrimitive2D.cpp
@@ -82,12 +82,12 @@ void KRPrimitive2D::fillQuad(const KRVector2D& p1, const KRVector2D& p2, const K
     };
     
     GLfloat colors[] = {
-        c1.r, c1.g, c1.b, c1.a,
-        c2.r, c2.g, c2.b, c2.a,
-        c3.r, c3.g, c3.b, c3.a,
-        c1.r, c1.g, c1.b, c1.a,
-        c3.r, c3.g, c3.b, c3.a,
-        c4.r, c4.g, c4.b, c4.a,
+        (GLfloat)c1.r, (GLfloat)c1.g, (GLfloat)c1.b, (GLfloat)c1.a,
+        (GLfloat)c2.r, (GLfloat)c2.g, (GLfloat)c2.b, (GLfloat)c2.a,
+        (GLfloat)c3.r, (GLfloat)c3.g, (GLfloat)c3.b, (GLfloat)c3.a,
+        (GLfloat)c1.r, (GLfloat)c1.g, (GLfloat)c1.b, (GLfloat)c1.a,
+        (GLfloat)c3.r, (GLfloat)c3.g, (GLfloat)c3.b, (GLfloat)c3.a,
+        (GLfloat)c4.r, (GLfloat)c4.g, (GLfloat)c4.b, (GLfloat)c4.a,
     };
     
     glVertexPointer(2, GL_SHORT, 0, vertices);
